Now Can Save lat and lng of reports but cannot render many heatmaps

+++ database/migrations/2020_10_04_073038_create_donations_table.php
@@ -16,7 +16,7 @@ class CreateDonationsTable extends Migration
         Schema::create('donations', function (Blueprint $table) {
             $table->id();
             $table->string('donation_email');
-            $table->unsignedBigInteger('user_id')->nullable();
+            $table->unsignedInteger('user_id')->nullable();
             $table->string('donation_name');
             $table->integer('donation_amount');
             $table->string('donation_transaction_id');
